# api-docs (1).json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8089","description":"Generated server url"}],"paths":{"/api/clients/{id}":{"get":{"tags":["client-rest-controller"],"operationId":"getClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientDTO"}}}}}},"put":{"tags":["client-rest-controller"],"operationId":"updateClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientDTO"}}}}}},"delete":{"tags":["client-rest-controller"],"operationId":"deleteClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/paiements":{"post":{"tags":["contrat-rest-controller"],"operationId":"savePaiement","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaiementDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaiementDTO"}}}}}}},"/api/contrats/sante":{"post":{"tags":["contrat-rest-controller"],"operationId":"saveSante","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContratSanteDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContratAssuranceDTO"}}}}}}},"/api/contrats/habitation":{"post":{"tags":["contrat-rest-controller"],"operationId":"saveHabitation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContratHabitationDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContratAssuranceDTO"}}}}}}},"/api/contrats/automobile":{"post":{"tags":["contrat-rest-controller"],"operationId":"saveAutomobile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContratAutomobileDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContratAssuranceDTO"}}}}}}},"/api/clients":{"get":{"tags":["client-rest-controller"],"operationId":"listClients","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClientDTO"}}}}}}},"post":{"tags":["client-rest-controller"],"operationId":"saveClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientDTO"}}}}}}},"/api/contrats/{id}/statut":{"patch":{"tags":["contrat-rest-controller"],"operationId":"updateStatut","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"statut","in":"query","required":true,"schema":{"type":"string","enum":["EN_COURS","VALIDE","RESILIE"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContratAssuranceDTO"}}}}}}},"/api/contrats":{"get":{"tags":["contrat-rest-controller"],"operationId":"listContrats","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ContratAssuranceDTO"}}}}}}}},"/api/contrats/{id}":{"get":{"tags":["contrat-rest-controller"],"operationId":"getContrat","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContratAssuranceDTO"}}}}}},"delete":{"tags":["contrat-rest-controller"],"operationId":"deleteContrat","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/contrats/{id}/paiements":{"get":{"tags":["contrat-rest-controller"],"operationId":"getPaiements","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaiementDTO"}}}}}}}},"/api/contrats/statut/{statut}":{"get":{"tags":["contrat-rest-controller"],"operationId":"getContratsByStatut","parameters":[{"name":"statut","in":"path","required":true,"schema":{"type":"string","enum":["EN_COURS","VALIDE","RESILIE"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ContratAssuranceDTO"}}}}}}}},"/api/clients/{id}/contrats":{"get":{"tags":["client-rest-controller"],"operationId":"getContratsClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ContratAssuranceDTO"}}}}}}}},"/api/clients/search":{"get":{"tags":["client-rest-controller"],"operationId":"searchClients","parameters":[{"name":"keyword","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClientDTO"}}}}}}}}},"components":{"schemas":{"ClientDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nom":{"type":"string"},"email":{"type":"string"}}},"PaiementDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date"},"montant":{"type":"number","format":"double"},"type":{"type":"string","enum":["MENSUALITE","PAIEMENT_ANNUEL","PAIEMENT_EXCEPTIONNEL"]},"contratId":{"type":"integer","format":"int64"}}},"ContratSanteDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dateSouscription":{"type":"string","format":"date"},"statut":{"type":"string","enum":["EN_COURS","VALIDE","RESILIE"]},"dateValidation":{"type":"string","format":"date"},"montantCotisation":{"type":"number","format":"double"},"dureeContrat":{"type":"integer","format":"int32"},"tauxCouverture":{"type":"number","format":"double"},"clientId":{"type":"integer","format":"int64"},"typeContrat":{"type":"string"},"niveauCouverture":{"type":"string","enum":["BASIQUE","INTERMEDIAIRE","PREMIUM"]},"nombrePersonnesCouvertes":{"type":"integer","format":"int32"}}},"ContratAssuranceDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dateSouscription":{"type":"string","format":"date"},"statut":{"type":"string","enum":["EN_COURS","VALIDE","RESILIE"]},"dateValidation":{"type":"string","format":"date"},"montantCotisation":{"type":"number","format":"double"},"dureeContrat":{"type":"integer","format":"int32"},"tauxCouverture":{"type":"number","format":"double"},"clientId":{"type":"integer","format":"int64"},"typeContrat":{"type":"string"}}},"ContratHabitationDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dateSouscription":{"type":"string","format":"date"},"statut":{"type":"string","enum":["EN_COURS","VALIDE","RESILIE"]},"dateValidation":{"type":"string","format":"date"},"montantCotisation":{"type":"number","format":"double"},"dureeContrat":{"type":"integer","format":"int32"},"tauxCouverture":{"type":"number","format":"double"},"clientId":{"type":"integer","format":"int64"},"typeContrat":{"type":"string"},"typeLogement":{"type":"string","enum":["APPARTEMENT","MAISON","LOCAL_COMMERCIAL"]},"adresse":{"type":"string"},"superficie":{"type":"number","format":"double"}}},"ContratAutomobileDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dateSouscription":{"type":"string","format":"date"},"statut":{"type":"string","enum":["EN_COURS","VALIDE","RESILIE"]},"dateValidation":{"type":"string","format":"date"},"montantCotisation":{"type":"number","format":"double"},"dureeContrat":{"type":"integer","format":"int32"},"tauxCouverture":{"type":"number","format":"double"},"clientId":{"type":"integer","format":"int64"},"typeContrat":{"type":"string"},"numeroImmatriculation":{"type":"string"},"marqueVehicule":{"type":"string"},"modeleVehicule":{"type":"string"}}}}}}